# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenSplit API",
        "description": "Enterprise backend API for OpenSplit - an open-source Splitwise alternative",
        "contact": {
            "name": "OpenSplit Team",
            "email": "support@opensplit.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/expenses": {
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Create a new expense",
                "description": "Creates a new expense with splits. The sum of all owed_share values must equal the total expense amount.",
                "operationId": "createExpense",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "amount",
                                    "group_id",
                                    "paid_by",
                                    "splits"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Dinner at restaurant"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01,
                                        "example": 300
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "paid_by": {
                                        "description": "User ID who paid",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "currency_code": {
                                        "type": "string",
                                        "maxLength": 3,
                                        "example": "USD"
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-01-03"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "splits": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "user_id",
                                                "paid_share",
                                                "owed_share"
                                            ],
                                            "properties": {
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "paid_share": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "minimum": 0,
                                                    "example": 300
                                                },
                                                "owed_share": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "minimum": 0,
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expense created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or split sum mismatch",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/expenses/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get expense by ID",
                "description": "Returns a single expense with its splits",
                "operationId": "getExpense",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Expense ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Expense"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expense not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Delete an expense",
                "description": "Deletes an expense and all its splits (cascade)",
                "operationId": "deleteExpense",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Expense ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expense deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expense not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/groups/{groupId}/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get all expenses for a group",
                "description": "Returns all expenses belonging to a specific group",
                "operationId": "getGroupExpenses",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "Group ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Expense"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ExpenseSplit": {
                "required": [
                    "user_id",
                    "paid_share",
                    "owed_share"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "paid_share": {
                        "type": "string",
                        "example": "100.00"
                    },
                    "owed_share": {
                        "type": "string",
                        "example": "33.33"
                    }
                },
                "type": "object"
            },
            "Expense": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "Dinner at restaurant"
                    },
                    "amount": {
                        "type": "string",
                        "example": "300.00"
                    },
                    "currency_code": {
                        "type": "string",
                        "example": "USD"
                    },
                    "paid_by": {
                        "type": "integer",
                        "example": 1
                    },
                    "group_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "expense_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-01-03"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "splits": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ExpenseSplit"
                        }
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}